% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvpa-package.R
\docType{package}
\name{bvpa-package}
\alias{bvpa-package}
\alias{bvpa}
\title{Bivariate Pareto Distribution}
\description{
Implements the EM algorithm with one-step Gradient Descent method to estimate
the parameters of the Block-Basu bivariate Pareto distribution with location
and scale. We also found parametric bootstrap and asymptotic confidence
intervals based on the observed Fisher information scale, shape parameters,
and exact confidence intervals for location parameters.
}
\references{
Biplab Paul and Arabin Kumar Dey (2023). An EM algorithm for absolutely
continuous Marshall-Olkin bivariate Pareto distribution with location
and scale, Preprint.

E L Lehmann and George Casella (1998). Theory of Point Estimation,
Springer, New York, doi.org/10.1007/b98854.

Bradley Efron and R J Tibshirani (1994). An Introduction to the Bootstrap, CRC press,
New York, doi.org/10.1201/9780429246593.

A P Dempster, N M Laird and D B Rubin (1977). Maximum Likelihood from
Incomplete Data via the EM Algorithm, Journal of the royal statistical society:
series B (methodological),
www.jstor.org/stable/2984875.
}
\author{
Biplab Paul <paul.biplab497@gmail.com> and Arabin Kumar Dey <arabin@iitg.ac.in>
}
\keyword{package}
