% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLf1.R
\name{mLf1}
\alias{mLf1}
\title{Marginal log-liklihood function of variable X1}
\usage{
mLf1(I, mu1, s1, a0, a1, a2)
}
\arguments{
\item{I}{baivariate observations.}

\item{mu1}{value of \eqn{\mu_1}.}

\item{s1}{value of \eqn{\sigma_1}.}

\item{a0}{value of \eqn{\alpha_0}.}

\item{a1}{value of \eqn{\alpha_1}.}

\item{a2}{value of \eqn{\alpha_2}.}
}
\value{
A scalar numeric, the marginal log-liklihood value of variable \eqn{X_1}.
}
\description{
Return the marginal log-liklihood value of variable \eqn{X_1}.
}
\examples{
dat <- rbb.bvpa(500, 0.1, 0.1, 0.8, 0.8, 2.0, 0.4, 0.5)
mLf1(dat, 0.1, 0.8, 2.0, 0.4, 0.5)

}
\author{
Biplab Paul <paul.biplab497@gmail.com> and Arabin Kumar Dey <arabin@iitg.ac.in>
}
