% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\docType{data}
\name{precipitation}
\alias{precipitation}
\title{Precipitation data}
\format{
A data frame with 36524 rows and 2 columns and the following variables:
\describe{
  \item{Time}{1st column represents Day.}
  \item{Precipitation..mm.}{2nd column represents daily accumulated precipitation (in mm)
   of the day.}
}
}
\source{
<https://www.polar.se/stoed-till-polarforskning/abisko-naturvetenskapliga-station/>
}
\usage{
data(precipitation)
}
\description{
The dataset contains daily accumulated precipitation data (in mm) from Abisko
Scientific Research Station in northern Sweden for 100 years, from 1st January
1913 to 31st December 2012.
}
\examples{
data(precipitation)
}
\keyword{datasets}
