% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul.R
\name{rbb.bvpa}
\alias{rbb.bvpa}
\title{Simulate from a Block-Basu Bivariate Pareto (BBBVPA) distribution}
\usage{
rbb.bvpa(n, mu1, mu2, sig1, sig2, alp0, alp1, alp2)
}
\arguments{
\item{n}{number of observations.}

\item{mu1}{value of \eqn{\mu_1}}

\item{mu2}{value of \eqn{\mu_2}}

\item{sig1}{value of \eqn{\sigma_1}}

\item{sig2}{value of \eqn{\sigma_2}}

\item{alp0}{value of \eqn{\alpha_0}}

\item{alp1}{value of \eqn{\alpha_1}}

\item{alp2}{value of \eqn{\alpha_2}}
}
\value{
numeric matrix.
}
\description{
Produces one or more samples from the specified BBBVPA distribution.
}
\examples{
cor(rbb.bvpa(500, 0.1, 0.1, 0.8, 0.8, 2.0, 0.4, 0.5))

}
\author{
Biplab Paul <paul.biplab497@gmail.com> and Arabin Kumar Dey <arabin@iitg.ac.in>
}
